move https_client into separate file
